+++ src/data.ts
@@ -1,4 +1,5 @@
-import { Position, Range, TextDocument, window } from 'vscode';
+import { Position, Range, TextDocument, Uri, window, workspace } from 'vscode';
+import { getContextFiles, removeContextFile } from './utility';
 
 export class CodeCompletions {
   completions: [string, string][] = [];
@@ -84,7 +85,6 @@ export class PromptBuilder {
 
   private activeFile: string;
   private activeFilePath: string;
-  private visibleFiles: { content: string; path: string }[];
   private lineEnding: string | null;
   private isSingleLineCompletions: boolean;
 
@@ -96,8 +96,6 @@ export class PromptBuilder {
     const lineInfo = this.createLineEnding(document, position);
     this.lineEnding = lineInfo.lineEnding;
     this.isSingleLineCompletions = lineInfo.isSingleLineCompletion;
-
-    this.visibleFiles = this.getVisibleFiles();
   }
 
   private createLineEnding(document: TextDocument, position: Position) {
@@ -137,6 +135,26 @@ export class PromptBuilder {
     return visibleFiles;
   }
 
+  private async getSelectedFiles() {
+    const contextFiles: string[] = getContextFiles();
+
+    return (
+      await Promise.all(
+        contextFiles.map(async (path) => {
+          try {
+            return {
+              path,
+              content: await workspace.fs.readFile(Uri.file(path)),
+            };
+          } catch (e) {
+            removeContextFile(path);
+            return { path, content: '' };
+          }
+        })
+      )
+    ).filter((x) => x.content);
+  }
+
   getFileInfo() {
     return {
       activeFile: this.activeFile,
@@ -145,8 +163,11 @@ export class PromptBuilder {
     };
   }
 
-  getPrompt() {
-    const files = this.visibleFiles;
+  async getPrompt() {
+    const visibleFiles = this.getVisibleFiles();
+    const selectedFiles = await this.getSelectedFiles();
+
+    const files = [...selectedFiles, ...visibleFiles];
     files.push({ content: this.activeFile, path: this.activeFilePath });
 
     return files