+++ src/extension.ts
@@ -3,6 +3,7 @@
 import * as vscode from 'vscode';
 import { LLMCompletionProvider } from './completion_provider';
 import { commands } from './commands';
+import { ContextSelectionView } from './ui/context_view';
 
 // This method is called when your extension is activated
 // Your extension is activated the very first time the command is executed
@@ -13,7 +14,16 @@ export function activate(context: vscode.ExtensionContext) {
   console.log('Version: ', context.extension.packageJSON['version']);
   LLMCompletionProvider.build();
 
+  vscode.commands.executeCommand(
+    'setContext',
+    'localcompletion:useContextGitignore',
+    vscode.workspace
+      .getConfiguration('localcompletion')
+      .get<boolean>('context_gitignore', true)
+  );
+
   context.subscriptions.push(
+    ContextSelectionView.build(context),
     vscode.languages.registerInlineCompletionItemProvider(
       { pattern: '**' },
       LLMCompletionProvider.instance()