+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Change Log
 
+## [Unreleased]
+
+### Added
+
+- New view for selecting additional context
+  - Optionally files, mentioned in the .gitignore are automatically excluded from the context selection
+
 ## [0.1.4]
 
 - Add status bar item with feed back when completion is ongoing or deactivated